move files around to create a better structure
