+++ tests/test_fairlearn_smoke.py
@@ -3,8 +3,8 @@
 
 import numpy as np
 import pandas as pd
-import fairlearn.moments as moments
-import fairlearn.classred as red
+import fairlearn.expgrad.moments as moments
+import fairlearn.expgrad.classred as red
 
 import pytest
